chore: move fluid texture & color querying methods, make public
Out of SelectFluidScreen, into ClientUtils. Useful for FTB Quests fluid
icon rendering.
