+++ fabric/src/main/java/dev/ftb/mods/ftblibrary/util/client/fabric/ClientUtilsImpl.java
@@ -1,0 +1,41 @@
+package dev.ftb.mods.ftblibrary.util.client.fabric;
+
+import dev.architectury.fluid.FluidStack;
+import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandlerRegistry;
+import net.minecraft.client.Minecraft;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
+
+public class ClientUtilsImpl {
+    public static ResourceLocation getStillTexture(FluidStack stack) {
+        var fluid = stack.getFluid();
+        var handler = FluidRenderHandlerRegistry.INSTANCE.get(fluid);
+
+        if (handler == null) {
+            return null;
+        }
+
+        var sprites = handler.getFluidSprites(Minecraft.getInstance().level, Minecraft.getInstance().level == null ? null : BlockPos.ZERO, fluid.defaultFluidState());
+        if (sprites[0] == null) {
+            return null;
+        }
+        return sprites[0].contents().name();
+    }
+
+
+    public static int getFluidColor(FluidStack stack) {
+        var fluid = stack.getFluid();
+        var handler = FluidRenderHandlerRegistry.INSTANCE.get(fluid);
+
+        var color = -1;
+        if (handler != null) {
+            if (Minecraft.getInstance().level != null) {
+                color = handler.getFluidColor(Minecraft.getInstance().level, BlockPos.ZERO, fluid.defaultFluidState());
+            } else {
+                color = handler.getFluidColor(null, null, fluid.defaultFluidState());
+            }
+        }
+
+        return color;
+    }
+}
